Extract common Chef-related sandbox code to Chef::CommonSandbox.
For the moment, the specs remain in chef_base_spec.rb and continue to
exercise this internal class.
